refactor: extract inbox implementation into own module
