+++ frappe/email/__init__.py
@@ -2,6 +2,7 @@
 # License: MIT. See LICENSE
 
 import frappe
+from frappe.utils.deprecations import deprecated
 
 
 def sendmail_to_system_managers(subject, content):
@@ -55,45 +56,19 @@ def get_system_managers():
 
 
 @frappe.whitelist()
+@deprecated
 def relink(name, reference_doctype=None, reference_name=None):
-	frappe.db.sql(
-		"""update
-			`tabCommunication`
-		set
-			reference_doctype = %s,
-			reference_name = %s,
-			status = "Linked"
-		where
-			communication_type = "Communication" and
-			name = %s""",
-		(reference_doctype, reference_name, name),
-	)
+	from frappe.communications.inbox import relink
+
+	return relink(name, reference_doctype, reference_name)
 
 
 @frappe.whitelist()
 @frappe.validate_and_sanitize_search_inputs
 def get_communication_doctype(doctype, txt, searchfield, start, page_len, filters):
-	user_perms = frappe.utils.user.UserPermissions(frappe.session.user)
-	user_perms.build_permissions()
-	can_read = user_perms.can_read
-	from frappe.modules import load_doctype_module
+	from frappe.communications.inbox import get_communication_doctype
 
-	com_doctypes = []
-	if len(txt) < 2:
-		for name in frappe.get_hooks("communication_doctypes"):
-			try:
-				module = load_doctype_module(name, suffix="_dashboard")
-				if hasattr(module, "get_data"):
-					for i in module.get_data()["transactions"]:
-						com_doctypes += i["items"]
-			except ImportError:
-				pass
-	else:
-		com_doctypes = [
-			d[0] for d in frappe.db.get_values("DocType", {"issingle": 0, "istable": 0, "hide_toolbar": 0})
-		]
-
-	return [[dt] for dt in com_doctypes if txt.lower().replace("%", "") in dt.lower() and dt in can_read]
+	return get_communication_doctype(doctype, txt, searchfield, start, page_len, filters)
 
 
 def get_cached_contacts(txt):
